+++ src/api/deprecated.rs
@@ -1343,8 +1343,7 @@ pub mod deprecated_array_functions {
         array: &mut Array,
         comparer: &str,
     ) -> RhaiResultOf<()> {
-        sort_by(ctx, array, FnPtr::new(comparer)?);
-        Ok(())
+        sort_by(ctx, array, FnPtr::new(comparer)?)
     }
     /// Remove all elements in the array that returns `true` when applied a function named by `filter`
     /// and return them as a new array.